[Refactor] mime 타입과 파일 시그니처 필터링 기능을 개선 - 기존의 enum의 values() 메서드를 사용했던 부분을 Map으로 변경하여 성능을 개선 - 파일 시그니처와 mime 타입의 필터링 기능을 Map에서 가져온 데이터를 비교하는 방식으로 변경
